Beginning implementation of Office panel
Clicking on a sub-button in a category will now load its corresponding page in the panel box. Currently only the "Center Information" page is done since I was just trying to figure out the most effecient way I could set it up, but it should be easy to do the rest now.