Implemented my code for Scorer package.
Refactored "addBall" method in BowlingGame to "bowled"
Refactores BowlingSession.onBowlEvent() to significantly reduce the workload instead letting BowlingGame handle it.
Added "mFinishedFrame" variable to BowlingGame to indicate when a frame is finished to let session know.
Removed currentBall and framesLeft from BowlingSession.ru

Added a variable mSessionFinished to indicate when all games are finished by all players. I moved the for loop inside NumberedSession.run() to a method newGame() inside NumberedSession and made that an abstract method in BowlingSession.

I also added refreshDisplay to NumberedSession.onBowlEvent()

run in NumberedSession inside Scorer package.

Inside Scorer.BowlingSession (approx line 99) is where we have a player bowl and then finished a frame. I think it is here that a credit should be subtracted and I believe this is how Vector does it as well. Vector doesn't care what frame you're on, once you're out of frames, you are done. As we discussed in the past, there are several reasons why it will be advantageous to have credits by frame instead of by the game. There are times when people only want to bowl half a game, people may be added later halfway into a game (we do want customers to be able to do that, but by the game we have to account for such exceptions which sounds messy), etc.

Added updateTo() to new BowlingGame